requirements_web.txt
flask>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
PyYAML>=6.0
numpy>=1.24.0
python-dotenv>=1.0.0